Simplified postgres by removing superflous interface
